+++ react/typings/vtex.styleguide.d.ts
@@ -1,3 +1,4 @@
 declare module 'vtex.styleguide' {
   export const Dropdown: any
+  export const Modal: any
 }